pyproject.toml
[tool.poetry]
name = "thetadata"
version = "0.9.13"
description = "A wrapper around Thetadata's api."
authors = ["David Mckim <davidmckim@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.10.2"
pandas = "^2.2.3"
httpx = "^0.27.2"
tenacity = "^9.0.0"
psutil = "^6.1.0"
wget = "^3.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.0"
pytest = "^8.3.3"
python-dotenv = "^1.0.1"
coverage = "^7.6.9"
selenium = "^4.27.1"
pytest-timeout = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
